requirements.txt
aiogram==2.25.2
Pillow==10.4.0
python-dotenv==1.0.1
